@@ -3,26 +3,26 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.4.0",
-    "@types/node": "^16.11.21",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
     "axios": "^0.25.0",
-    "bootstrap": "^5.1.3",
-    "bootstrap-icons": "^1.8.1",
-    "echarts": "^5.2.2",
-    "echarts-for-react": "^3.0.2",
-    "fs": "0.0.1-security",
+    "bootstrap": "^5.2.3",
+    "bootstrap-icons": "^1.10.3",
+    "budget-view-chart": "^0.1.2",
     "react": "^17.0.2",
-    "react-bootstrap": "^2.1.2",
+    "react-bootstrap": "^2.7.2",
     "react-dom": "^17.0.2",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "5.0.0",
-    "typescript": "^4.5.5",
-    "web-vitals": "^2.1.3"
+    "react-router-dom": "^6.10.0",
+    "web-vitals": "^2.1.4"
+  },
+  "devDependencies": {
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^12.1.5",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^29.5.0",
+    "@types/node": "^16.18.23",
+    "@types/react": "^17.0.55",
+    "@types/react-dom": "^17.0.19",
+    "react-scripts": "^5.0.1",
+    "typescript": "^5.0.3"
   },
   "scripts": {
     "start": "react-scripts start",
